requirements.txt
# Core async / web
httpx==0.27.0

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.2.2

# Browser automation
playwright==1.43.0
Appium-Python-Client>=4.0,<6.0

# Data handling (optional but safe)
pydantic==2.7.1

# Typing / utils (usually builtin but safe for compatibility)
typing-extensions==4.12.2

#run the line
#playwright install
#run: python scripts/run_pipeline.py https://3afsatunisie.com


python-dotenv
requests
openpyxl
pillow
